requirements.txt
pandas==2.1.4
flask==3.0.0
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0 